requirements.txt
# 基础依赖
colorama>=0.4.6
neo4j>=5.8.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
pickle5>=0.0.12; python_version < '3.8'
jieba>=0.42.1
# Web界面依赖
flask>=2.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
yaml
PyYAML>=6.0.1
# 类型提示
typing-extensions>=4.7.0
faiss-cpu
# 新增依赖
tqdm>=4.62.0
tiktoken>=0.5.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pyyaml>=6.0
psutil>=5.9.0